[level 0] Title: 두 수의 나눗셈, Time: 0.01 ms, Memory: 4.02 MB -BaekjoonHub

+++ 프로그래머스/0/120806. 두 수의 나눗셈/두 수의 나눗셈.cpp
@@ -5,6 +5,6 @@ using namespace std;
 
 int solution(int num1, int num2) {
     float temp = (float)num1 / num2 * 1000;
-    int answer =  (int)temp;
+    int answer =  temp;
     return answer;
 }
